Added username string, setting username tests, and moved helper functions from own file to own object that is part of the main component file
