+++ src/components/Navbar/Navbar.js
@@ -2,8 +2,27 @@ import * as React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { GithubLoginButton } from "react-social-login-buttons";
 import './Navbar.css';
-import {login, signOut} from "./helpers";
 import { FaSearch } from 'react-icons/fa';
+
+import { firebaseAuth } from "../../utility/firebaseFascade";
+import {authActionCreators} from "../../actions";
+
+export const helpers = {
+  login: async (dispatch) => {
+    const startLoadingAction = authActionCreators.startLoading();
+    dispatch(startLoadingAction);
+    await firebaseAuth.loginWithGithub();
+    const loginAction = authActionCreators.login();
+    dispatch(loginAction);
+    const userModel = await firebaseAuth.fetchOwnUserModel();
+    const storeUsernameAction = authActionCreators.finishLogin(userModel);
+    dispatch(storeUsernameAction);
+  },
+  signOut: async (dispatch) => {
+    await firebaseAuth.signOut();
+    dispatch(authActionCreators.signOut());
+  }
+};
 
 export const NavbarComponent = () => {
   const auth = useSelector( state => state.auth );
@@ -18,12 +37,12 @@ export const NavbarComponent = () => {
 
 
     {/*Logged out nav bar items*/}
-    {!loggedIn && <GithubLoginButton id="githubButton" onClick={()=>login(dispatch)}/>}
+    {!loggedIn && <GithubLoginButton id="githubButton" onClick={()=>helpers.login(dispatch)}/>}
 
     {/*Logged in nav bar items*/}
     {!!loggedIn &&
       <div className="navbarTextContainer">
-        <div className="navBarTextItem" id="signOut" onClick={()=>signOut(dispatch)}>
+        <div className="navBarTextItem" id="signOut" onClick={()=>helpers.signOut(dispatch)}>
           Logout
         </div>
         <div className="navBarTextItem">
