+++ src/components/Navbar/Navbar.test.js
@@ -2,8 +2,7 @@ import React from 'react';
 import configureMockStore from 'redux-mock-store';
 import Adapter from "enzyme-adapter-react-16/build";
 import {configure, mount, shallow} from 'enzyme';
-import Navbar from './Navbar';
-import * as helpers from './helpers';
+import Navbar, {helpers} from './Navbar';
 import {Provider} from "react-redux";
 import {firebaseAuth} from "../../utility/firebaseFascade";
 import {authActionCreators, authActions} from "../../actions";
