+++ src/components/Register/Register.css
@@ -26,6 +26,7 @@
   border-radius: 0.25rem;
   border-color: transparent;
   box-shadow: 0 0 0 0.1rem #9d9d9d;
+  margin-top: 2rem;
 }
 .formContainer input:focus {
   box-shadow: 0 0 0 0.25rem #91b9ff;
